+++ msgpack-core/src/main/java/org/msgpack/core/annotations/Nullable.java
@@ -13,11 +13,11 @@
 //    See the License for the specific language governing permissions and
 //    limitations under the License.
 //
-package org.msgpack.core;
+package org.msgpack.core.annotations;
 
 /**
  * Annotates a field which can be null
  */
-@interface Nullable
+public @interface Nullable
 {
 }
